bug: fix resume issue
When an exercise is resume, audio also resumes even if it was not playing when the exercise was paused. The issue was caused because the value of `audioStatus` persists between re-renders, so every exercise after the first one started with its value being `AudioStatus.playing`.

Note: This doesn't occur when manually going to the next or previous exercise because these involve resetting the variable